+++ web/src/components/hooks/index.js
@@ -1,11 +1,8 @@
 import { UseIsHelmManaged, useIsHelmManaged } from "./useIsHelmManaged";
 import { useSaveConfig } from "./useSaveConfig";
-import { useDownloadValues, UseDownloadValues } from "./useDownloadValues";
 
 export {
   UseIsHelmManaged,
-  useDownloadValues,
-  UseDownloadValues,
   useIsHelmManaged,
   useSaveConfig,
 };
