+++ web/src/features/AppVersionHistory/api/index.tsx
@@ -1,0 +1,3 @@
+import { useDownloadValues } from "./useDownloadValues";
+
+export { useDownloadValues };
